Issue 203: Merge MediaFileEdit and MediaFileAdd
