+++ app/src/main/resources/ApplicationResources.properties
@@ -422,7 +422,7 @@ error.add.user.badUserName=Invalid user name (must be alpha-numerics only).
 error.add.user.missingPassword=You must specify a password.
 error.upload.dirmax=You cannot exceed the maximum directory size of {0} MB.
 error.upload.disabled=File Upload has been turned off
-error.upload.file=No file selected
+error.upload.nofile=No file selected
 error.upload.filemax=File <b>{0}</b> larger than maximum <b>{1}</b> MB allowed
 error.upload.forbiddenFile=File <b>{0}</b> content-type <b>{1}</b> not allowed
 error.general=ERROR: Unexpected Exception [{0}] has been logged.
@@ -734,8 +734,6 @@ mediaFileAdd.anchor=Target URL
 mediaFileAdd.anchor.tooltip=If the image is clicked on, webpage URL (including http:// or \
   https://) to redirect to.
 
-mediaFileAdd.notes.tooltip=Any notes you wish to make about this media file.  Not intended to be displayed publicly.
-
 mediaFileAdd.fileLocation=Choose file to upload
 mediaFileAdd.copyright=Copyright statement
 mediaFileAdd.directory=Folder
@@ -787,7 +785,7 @@ mediaFileEdit.fileDimensions=<b>Image</b>: {0} X {1} pixels
 mediaFileEdit.includeGallery=Include in Gallery?
 mediaFileEdit.includeGalleryHelp=Check this box if you would like this media \
 file included in the new media file RSS and Atom feeds for your weblog.
-mediaFileEdit.updateFileContents=Update File Contents
+mediaFileEdit.fileLocation=Update File Contents
 mediaFileEdit.clickToView=Click to view in new window
 mediaFileEdit.permalink=Link (URL)
 mediaFileEdit.linkTitle=Link to Media File, opens in new window
