+++ app/src/main/resources/ApplicationResources_de.properties
@@ -334,7 +334,7 @@ error.untranslated={0}
 #---------------------------------------------------------------- Error messages
 error.upload.dirmax=Die maximale Verzeichnisgr\u00F6\u00DFe von {0} MB kann nicht \u00FCberschritten werden.
 error.upload.disabled=Die Funktion 'Datei hochladen' wurde deaktiviert
-error.upload.file=Es wurde keine Datei ausgew\u00E4hlt
+error.upload.nofile=Es wurde keine Datei ausgew\u00E4hlt
 error.upload.filemax=Die Datei \u00FCberschreitet die maximal erlaubte Gr\u00F6\u00DFe von {0} MB
 error.upload.forbiddenFile=Es k\u00F6nnen nur Dateien hochgeladen werden, die in {0} enden
 errorPage.message=Roller hat eine unerwartete Ausnahmebedingung festgestellt und aufgezeichnet.
