+++ app/src/main/resources/ApplicationResources_es.properties
@@ -182,7 +182,7 @@ error.add.user.badUserName=Nombre de usuario no v\u00E1lido (debe ser s\u00F3lo 
 error.add.user.missingPassword=Debe especificar una contrase\u00F1a
 error.upload.dirmax=No se puede exceder el tama\u00F1o m\u00E1ximo de directorio de {0} MB.
 error.upload.disabled=La carga de ficheros ha sido deshabilitada
-error.upload.file=Ning\u00FAn fichero seleccionado
+error.upload.nofile=Ning\u00FAn fichero seleccionado
 error.upload.filemax=Fichero mayor que el m\u00E1ximo permitido\: {0} MB
 error.upload.forbiddenFile=Puede cargar s\u00F3lo ficheros acabados en {0}
 error.general=ERROR\: Excepci\u00F3n inesperada [{0}] ha sido registrada.
