+++ app/src/main/resources/ApplicationResources_fr.properties
@@ -338,7 +338,7 @@ error.add.user.missingPassword=Vous devez entrer un mot de passe.
 
 error.upload.dirmax=Vous ne pouvez pas dépasser la limite de {0} MB pour la taille d'un répertoire.
 error.upload.disabled=Le chargement de fichier est désactivé
-error.upload.file=Aucun fichier sélectionné
+error.upload.nofile=Aucun fichier sélectionné
 error.upload.filemax=Ce fichier dépasse la taille autoriséee de {0} MB
 error.upload.forbiddenFile=Vous ne pouvez charger que les fichiers avec l'une des extensions suivantes : {0}
 error.general=ERREUR : une erreur inattendue [{0}] est survenue.
