+++ app/src/main/resources/ApplicationResources_ja.properties
@@ -275,7 +275,7 @@ error.add.user.missingPassword=\u5BFE\u5FDC\u3059\u308B\u30D1\u30B9\u30EF\u30FC\
 
 error.upload.dirmax=\u3053\u306E\u30C7\u30A3\u30EC\u30AF\u30C8\u30EA\u306B{0} MB\u4EE5\u4E0A\u306E\u30A2\u30C3\u30D7\u30ED\u30FC\u30C9\u306F\u3067\u304D\u307E\u305B\u3093
 error.upload.disabled=\u30A2\u30C3\u30D7\u30ED\u30FC\u30C9\u306F\u8A31\u53EF\u3055\u308C\u3066\u3044\u307E\u305B\u3093
-error.upload.file=\u30D5\u30A1\u30A4\u30EB\u304C\u6307\u5B9A\u3055\u308C\u3066\u3044\u307E\u305B\u3093
+error.upload.nofile=\u30D5\u30A1\u30A4\u30EB\u304C\u6307\u5B9A\u3055\u308C\u3066\u3044\u307E\u305B\u3093
 error.upload.filemax=1\u30D5\u30A1\u30A4\u30EB\u4E0A\u9650{0} MB\u3092\u8D8A\u3048\u3066\u3044\u307E\u3059
 error.upload.forbiddenFile=\u5F62\u5F0F{0}\u306E\u30A2\u30C3\u30D7\u30ED\u30FC\u30C9\u306F\u8A31\u53EF\u3055\u308C\u3066\u3044\u307E\u305B\u3093
 
@@ -1312,7 +1312,7 @@ pageForm.action=\u30C6\u30F3\u30D7\u30EC\u30FC\u30C8\u306E\u7DE8\u96C6
 mediaFileSuccess.size=\u30B5\u30A4\u30BA\:
 configForm.allowUserWeblogCreation=\u30E6\u30FC\u30B6\u306B\u65B0\u3057\u3044\u30D6\u30ED\u30B0\u306E\u4F5C\u6210\u3092\u8A31\u53EF\u3057\u307E\u3059\u304B?
 categoryForm.requiredFields={0} \u306F\u5FC5\u9808\u9805\u76EE\u3067\u3059
-mediaFileEdit.updateFileContents=\u30D5\u30A1\u30A4\u30EB\u306E\u5185\u5BB9\u3092\u66F4\u65B0\u3059\u308B
+mediaFileEdit.fileLocation=\u30D5\u30A1\u30A4\u30EB\u306E\u5185\u5BB9\u3092\u66F4\u65B0\u3059\u308B
 websiteSettings.error.processingBlacklist=\u30D6\u30E9\u30C3\u30AF\u30EA\u30B9\u30C8\u306E\u51E6\u7406\u30A8\u30E9\u30FC\: {0}
 mediaFileView.ge=>\=
 weblogEdit.mediaCastUrlMalformed=\u30A8\u30F3\u30AF\u30ED\u30FC\u30B8\u30E3URL\u304C\u4E0D\u6B63\u3067\u3059\u3002
