+++ app/src/main/resources/ApplicationResources_ko.properties
@@ -363,7 +363,7 @@ error.add.user.missingPassword=\ube44\ubc00\ubc88\ud638\ub97c \uc124\uc815\ud558
 
 error.upload.dirmax=\ub514\ub809\ud130\ub9ac \ud06c\uae30\ub294 {0} MB\ub97c \ucd08\uacfc\ud560 \uc218 \uc5c6\uc2b5\ub2c8\ub2e4.
 error.upload.disabled=\ud30c\uc77c \uc5c5\ub85c\ub4dc \uae30\ub2a5\uc774 \ud65c\uc131\ud654\ub418\uc5b4 \uc788\uc9c0 \uc54a\uc2b5\ub2c8\ub2e4.
-error.upload.file=\uc120\ud0dd\ub41c \ud30c\uc77c\uc774 \uc5c6\uc2b5\ub2c8\ub2e4.
+error.upload.nofile=\uc120\ud0dd\ub41c \ud30c\uc77c\uc774 \uc5c6\uc2b5\ub2c8\ub2e4.
 error.upload.filemax=\ud5c8\uc6a9\ub41c \ucd5c\ub300 \ud30c\uc77c \ud06c\uae30\ub97c \ucd08\uacfc\ud569\ub2c8\ub2e4: {0} MB
 error.upload.forbiddenFile=\ud5c8\uc6a9\ub41c \ud655\uc7a5\uc790\uc640 \ud5c8\uc6a9\ub41c \ucee8\ud150\ud2b8 \uc720\ud615\uc758 \ud30c\uc77c\ub9cc \uc5c5\ub85c\ub4dc \uac00\ub2a5\ud569\ub2c8\ub2e4: {0}
