+++ app/src/main/resources/ApplicationResources_zh_CN.properties
@@ -199,7 +199,7 @@ error.trackbackNotAllowed=\u53D1\u9001\u53CD\u5411\u8DDF\u8E2A\u5931\u8D25\u3002
 error.untranslated={0}
 error.upload.dirmax=\u60A8\u4E0D\u80FD\u8D85\u8FC7\u76EE\u5F55\u7684\u6700\u5927\u9650\u989D\uFF1A {0} MB\u3002
 error.upload.disabled=\u6587\u4EF6\u4E0A\u4F20\u529F\u80FD\u88AB\u7981\u7528
-error.upload.file=\u5199\u5165\u6587\u4EF6\u5931\u8D25
+error.upload.nofile=\u5199\u5165\u6587\u4EF6\u5931\u8D25
 error.upload.filemax=\u60A8\u4E0D\u80FD\u8D85\u8FC7\u5355\u4E2A\u6587\u4EF6\u7684\u6700\u5927\u9650\u989D\uFF1A {0} MB
 error.upload.forbiddenFile=\u60A8\u53EA\u80FD\u4E0A\u4F20\u4EE5 {0} \u7ED3\u5C3E\u7684\u6587\u4EF6
 errorPage.message=Roller\u9047\u5230\u4E00\u4E2A\u610F\u5916\u9519\u8BEF\u3002
